​​refactor: 영화 포스터 이미지 컴포넌트 분리
